Added card target selection.
Created prefabs for battleground gameobjects.
Added "Require Target" field in the card editor.  This option will later become a section with more targeting details.
Updated existing card data to accomodate the "Require Target" attribute.
Cards that require a target will now display a targeting arrow for the player to select a target on the battleground. This arrow will appear when a card is dragged from the hand, and disappear when the card is dropped or dragged back to the hand area.

Added support for more resolutions.
Improved the hand resize and card spacing in the hand to fit standard resolutions.
Increased the resolution of card text again to match the resolution changes.