Added Character class
The Character class derives from BaseUnit and is for playable units.

Added a CharacterType field for cards, this is used for character exclusive cards.

Added OwnerID for cards, this is for the card to keep track of its BaseUnit owner.

Cards now have an Owner attribute, it isn't functional yet.